Move PushEvent to Push namespace #52
