+++ definitions2/DevboardLib/GitHubWebhook/Hook/Issue/IssueEvent.yaml
@@ -1,6 +1,6 @@
 definition:
   type: Interface
   instanceOf: DevboardLib\GitHubWebhook\Hook\Issue\IssueEvent
-  parent: DevboardLib\GitHubWebhook\Hook\PushEvent
+  parent: DevboardLib\GitHubWebhook\Hook\Push\PushEvent
   constants: {  }
   methods: {  }
