+++ spec/Hook/Push/DeleteBranchPushEventSpec.php
@@ -14,7 +14,7 @@ use DevboardLib\GitHubWebhook\Core\Repo;
 use DevboardLib\GitHubWebhook\Core\Sender;
 use DevboardLib\GitHubWebhook\Hook\GitHubHookEvent;
 use DevboardLib\GitHubWebhook\Hook\Push\DeleteBranchPushEvent;
-use DevboardLib\GitHubWebhook\Hook\PushEvent;
+use DevboardLib\GitHubWebhook\Hook\Push\PushEvent;
 use PhpSpec\ObjectBehavior;
 
 class DeleteBranchPushEventSpec extends ObjectBehavior
