+++ src/Hook/IssueComment/IssueCommentEvent.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace DevboardLib\GitHubWebhook\Hook\IssueComment;
 
-use DevboardLib\GitHubWebhook\Hook\PushEvent;
+use DevboardLib\GitHubWebhook\Hook\Push\PushEvent;
 
 interface IssueCommentEvent extends PushEvent
 {
