+++ src/Hook/Push/PushEvent.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace DevboardLib\GitHubWebhook\Hook\Push;
+
+use DevboardLib\GitHubWebhook\Hook\GitHubHookEvent;
+
+interface PushEvent extends GitHubHookEvent
+{
+}
