--- src/Hook/PushEvent.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace DevboardLib\GitHubWebhook\Hook;
-
-interface PushEvent extends GitHubHookEvent
-{
-}
